+++ client/tests/factories/revise-task.js
@@ -1,0 +1,9 @@
+import FactoryGuy from 'ember-data-factory-guy';
+
+FactoryGuy.define('revise-task', {
+  default: {
+    title: 'Revise Manuscript',
+    type: 'ReviseTask',
+    completed: false,
+  }
+});