+++ CHANGELOG.md
@@ -13,6 +13,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 * Invitations no longer enter their edit state by default
 * Updated URLs to expose the manuscript's short DOI.  Papers can now be referenced
   by /papers/JOURNAL.DOI .  The app was updated to use these as the preferred links.
+* A user can now mark a card as incomplete at any time when it is in an editable state
 ### Deprecated
 ### Removed
 ### Fixed